requirements.txt
asgiref==3.4.1
cachetools==4.2.4
certifi==2021.10.8
cfgv==3.4.0
charset-normalizer==2.0.9
click==8.0.3
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.8
Django==4.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.21.7
filelock==3.15.4
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
httplib2==0.20.2
identify==2.6.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mypy-extensions==0.4.3
nodeenv==1.9.1
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==24.1
pathspec==0.9.0
platformdirs==4.2.2
pre-commit==3.8.0
protobuf==3.19.1
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
pytz==2021.3
PyYAML==6.0.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
sentry-sdk==2.14.0
simplejson==3.19.3
six==1.16.0
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3
virtualenv==20.26.3
google-auth
fcm-django
firebase-admin